Fixed a bug in DateTimePatternGenerator and updated some test cases.  The fix allow us to replace most of yyyy in the pattern resources with y.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/6476


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@24465 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/dtptngen.cpp
@@ -1457,7 +1457,7 @@ DateTimeMatcher::set(const UnicodeString& pattern, FormatParser* fp, PtnSkeleton
         if ( row->type > 0) {
             subTypeValue += field.length();
         }
-        skeletonResult.type[typeValue] = (int8_t)subTypeValue;
+        skeletonResult.type[typeValue] = subTypeValue;
     }
     copyFrom(skeletonResult);
 }
